+++ config/webpack.config.prod.js
@@ -281,6 +281,21 @@ function appConfig() {
         minifyURLs: true
       }
     }),
+    new HtmlWebpackPlugin({
+      inject: 'head',
+      filename: paths.shireICalOutput,
+      template: paths.shireICalHtml,
+      minify: {
+        removeComments: true,
+        removeRedundantAttributes: true,
+        useShortDoctype: true,
+        removeEmptyAttributes: true,
+        removeStyleLinkTypeAttributes: true,
+        keepClosingSlash: true,
+        minifyCSS: true,
+        minifyURLs: true
+      }
+  }),
     // Makes some environment variables available to the JS code, for example:
     // if (process.env.NODE_ENV === 'production') { ... }. See `./env.js`.
     // It is absolutely essential that NODE_ENV was set to production here.